backend/requirements.txt
mysqlclient>=1.3.14
Django>=2.2.4
python-decouple>=3.1
djangorestframework>=3.10.2
djangorestframework-simplejwt>=3.3
django-cors-headers>=2.4.0
